Make it possible to exclude link from header nav
Currently if you don't pass a `href` the navigation item link won't render when
using the header macro. This PR changes the navigation link to render
without the link wrapper if `item.text` or `item.html` is set but `href`
isn't.

We would expect this to be mainly useful for rendering a form for instance
for adding a log out button. The team will need to style any enclosed form
elements or buttons correctly.